+++ neuronquant/data/forex.py
@@ -21,11 +21,12 @@ def forex_rates(user, password, pairs='', fmt='csv'):
 class ConnectTrueFX(object):
     def __init__(self, user=None, password=None, auth_file='default.json', pairs='', fmt='csv'):
         #NOTE Without authentification you still can acess some data
+        #FIXME Not authorized response prevent from downloading quotes. However later you indeed retrieve 10 defaults
         self._code = None
         if (user is None) or (password is None):
             log.info('No credentials provided, trying to read configuration file')
             try:
-                config   = json.load(open('/'.join([os.environ['QTRADE'], 'config', auth_file]), 'r'))['truefx']
+                config   = json.load(open('/'.join([os.path.expanduser('~/.quantrade'), auth_file]), 'r'))['truefx']
                 user     = config['user']
                 password = config['password']
             except: