+++ neuronquant/data/ziplinesources/loader.py
@@ -18,8 +18,10 @@ from collections import OrderedDict
 from datetime import datetime
 
 from zipline.data.treasuries import get_treasury_data
-from zipline.data.benchmarks import get_benchmark_returns
+#from zipline.data.benchmarks import get_benchmark_returns
 from zipline.protocol import DailyReturn
+
+import neuronquant.utils.datautils as datautils
 
 from operator import attrgetter
 
@@ -46,15 +48,22 @@ class LiveBenchmark(object):
 
         # Getting today benchmark return
         #NOTE Seems shit but later, previous days could be used to compute indicators
-        last_bench_return = get_benchmark_returns(bm_symbol, start_date=(event_dt - pd.datetools.Day(self.loopback)))
-        last_bench_return = last_bench_return[-1]
-        print('Benchmark on {}: {}'.format(last_bench_return.date, last_bench_return.returns))
+        #last_bench_return = get_benchmark_returns(bm_symbol, start_date=(event_dt - pd.datetools.Day(self.loopback)))
+        #last_bench_return = last_bench_return[-1]
+        #print('Benchmark on {}: {}'.format(last_bench_return.date, last_bench_return.returns))
+
+        #TODO More efficient way to navigate in a dit
+        for exchange in datautils.Exchange:
+            if datautils.Exchange[exchange]['index'] == bm_symbol:
+                code = datautils.Exchange[exchange]['code']
+                break
 
         bm_returns = []
         while event_dt < self.last_trading_day:
             #TODO Current value to give
             #TODO Append only if trading day and market hour
-            bm_returns.append(DailyReturn(date=event_dt.replace(microsecond=0), returns=last_bench_return.returns))
+            #bm_returns.append(DailyReturn(date=event_dt.replace(microsecond=0), returns=last_bench_return.returns))
+            bm_returns.append(DailyReturn(date=event_dt.replace(microsecond=0), returns=code))
             #TODO Frequency control
             event_dt += self.offset