+++ scripts/install.sh
@@ -3,32 +3,43 @@
 clear
 echo "-- QuanTrade installation script"
 
+if [ $(whoami) != 'root' ]; then
+    echo '** Error: This installation script needs root permissions'
+    exit 1
+fi
+
 echo "Environment setup"
 #TODO Add a test
 echo "Creating project configuraiton directory: ~/.quantrade"
 mkdir $HOME/.quantrade
 echo "Copying configuration file templates..."
 echo "Note: do not forget to custom them at the end of the script"
-mv ../config/tpl_local.sh $HOME/.quantrade/local.sh
-mv ../config/tpl_default.json $HOME/.quantrade/default.json
+cp ../config/tpl_local.sh $HOME/.quantrade/local.sh
+cp ../config/tpl_default.json $HOME/.quantrade/default.json
+cp ../config/tpl_shiny-server.config $HOME/.quantrade/shiny-server.config
 
 echo "Writting in ~/.bashrc for permanent automatic environment load"
 echo "source $HOME/.quantrade/local.sh" >> $HOME/.bashrc
 
 echo "Sourcing environment setup file."
-source ../config/local.sh
+source ~/.quantrade/local.sh
+
+echo "Restoring user permission"
+chown -R $USER ~/.quantrade
+chgrp -R $USER ~/.quantrade
 
 echo "installation of debian packages..."
+apt-get install git libzmq-dev r-base python2.7 mysql-server libmysqlclient-dev npm pip planner python-dev
 
 echo "Writing python quantrade dependances"
-cat ./qt_*.txt > qt_deps.txt
+cat ./scripts/qt_*.txt > qt_deps.txt
 echo "Installing dependances with pip..."
-sudo ./ordered_pip.sh qt_deps.txt
+./scripts/ordered_pip.sh qt_deps.txt
 
 echo "Writing python quantrade dependances"
-cat ./z_*.txt > z_deps.txt
+cat ./scripts/z_*.txt > z_deps.txt
 echo "Installing dependances with pip..."
-sudo ./ordered_pip.sh z_deps.txt
+./scripts/ordered_pip.sh z_deps.txt
 
 echo "R installation, assuming you already use it at least once"
 ./scripts/install_r_packages.R
@@ -39,7 +50,7 @@ wget http://nodejs.org/dist/v0.10.0/node-v0.10.0.tar.gz
 tar xvzf node-v0.10.0.tar.gz && cd node-v0.10.0
 ./configure
 make
-sudo make install
+make install
 
 echo "Node modules for QuanTrade installation, locally"
 cd $QTRADE/server